Applying for someone else
Created a space for prototyping for applying for someone else.

Archived the original work for parents applying on behalf of minors.

Removed the old flow for parents applying for minors as it has been superceeded by the latest work including physical injuries and financial losses.

Added some ideas for the 'reminder banner'